+++ plugins/org.eclipse.n4js.ui/src/org/eclipse/n4js/ui/N4JSExternalLibraryStartup.java
@@ -10,12 +10,12 @@
  */
 package org.eclipse.n4js.ui;
 
+import org.eclipse.n4js.external.GitCloneSupplier;
+import org.eclipse.n4js.external.NpmLogger;
+import org.eclipse.n4js.external.libraries.ExternalLibrariesActivator;
 import org.eclipse.ui.IStartup;
 
 import com.google.inject.Inject;
-
-import org.eclipse.n4js.external.GitCloneSupplier;
-import org.eclipse.n4js.external.libraries.ExternalLibrariesActivator;
 
 /**
  * N4JS IDE startup hook to makes sure that the local Git repository exists for the type definition files.
@@ -25,12 +25,18 @@ public class N4JSExternalLibraryStartup implements IStartup {
 	@Inject
 	private GitCloneSupplier gitCloneSupplier;
 
+	@Inject
+	private NpmLogger logger;
+
 	@Override
 	public void earlyStartup() {
 		// Client code can still clone the repository on demand. (Mind plug-in UI tests.)
 		if (ExternalLibrariesActivator.requiresInfrastructureForLibraryManager()) {
 			new Thread(() -> {
 				gitCloneSupplier.get();
+				if (!gitCloneSupplier.isSuccessfullyCloned()) {
+					logger.logInfo("Error during setting up the npm type definitions. See error log for details.");
+				}
 			}).start();
 		}
 	}
